backend/requirements.txt
flask
flask-cors
gunicorn
google-cloud-logging
google-cloud-firestore
firebase-admin
langchain-anthropic
langchain